Moved log data to LogStore
