+++ src/ContainerHome.react.js
@@ -90,7 +90,7 @@ var ContainerHome = React.createClass({
                 <ContainerHomePreview />
               </div>
               <div className="right">
-                <ContainerHomeLogs />
+                <ContainerHomeLogs/>
                 <ContainerHomeFolders container={this.props.container} />
               </div>
             </div>
@@ -116,7 +116,7 @@ var ContainerHome = React.createClass({
           <div className="details-panel home">
             <div className="content">
               <div className="left">
-                <ContainerHomeLogs />
+                <ContainerHomeLogs/>
               </div>
               {right}
             </div>
